refactor(app-check): remove deprecated namespaced API
