+++ packages/ai/lib/public-types.ts
@@ -17,7 +17,7 @@
 
 import { ReactNativeFirebase } from '@react-native-firebase/app';
 import { FirebaseAuthTypes } from '@react-native-firebase/auth';
-import { FirebaseAppCheckTypes } from '@react-native-firebase/app-check';
+import type { AppCheck } from '@react-native-firebase/app-check';
 import { Backend } from './backend';
 
 export * from './types';
@@ -40,7 +40,7 @@ export interface AIOptions {
    * Whether to use App Check limited use tokens. Defaults to false.
    */
   useLimitedUseAppCheckTokens?: boolean;
-  appCheck?: FirebaseAppCheckTypes.Module | null;
+  appCheck?: AppCheck | null;
   auth?: FirebaseAuthTypes.Module | null;
 }
 
@@ -90,7 +90,7 @@ export interface AI {
    * The {@link @firebase/app!FirebaseApp} this {@link AI} instance is associated with.
    */
   app: ReactNativeFirebase.FirebaseApp;
-  appCheck?: FirebaseAppCheckTypes.Module | null;
+  appCheck?: AppCheck | null;
   auth?: FirebaseAuthTypes.Module | null;
   /**
    * A {@link Backend} instance that specifies the configuration for the target backend,
