+++ packages/ai/lib/service.ts
@@ -18,19 +18,19 @@
 import { ReactNativeFirebase } from '@react-native-firebase/app';
 import { AI, Backend } from './public-types';
 import { FirebaseAuthTypes } from '@react-native-firebase/auth';
-import { FirebaseAppCheckTypes } from '@react-native-firebase/app-check';
+import type { AppCheck } from '@react-native-firebase/app-check';
 import { VertexAIBackend } from './backend';
 
 export class AIService implements AI {
   auth: FirebaseAuthTypes.Module | null;
-  appCheck: FirebaseAppCheckTypes.Module | null;
+  appCheck: AppCheck | null;
   location: string;
 
   constructor(
     public app: ReactNativeFirebase.FirebaseApp,
     public backend: Backend,
     auth?: FirebaseAuthTypes.Module,
-    appCheck?: FirebaseAppCheckTypes.Module,
+    appCheck?: AppCheck,
   ) {
     this.auth = auth || null;
     this.appCheck = appCheck || null;
