+++ packages/app-check/e2e/appcheck.e2e.js
@@ -82,332 +82,6 @@ function decodeJWT(token) {
 }
 
 describe('appCheck()', function () {
-  describe('firebase v8 compatibility', function () {
-    beforeEach(async function beforeEachTest() {
-      // @ts-ignore
-      globalThis.RNFB_SILENCE_MODULAR_DEPRECATION_WARNINGS = true;
-    });
-
-    afterEach(async function afterEachTest() {
-      // @ts-ignore
-      globalThis.RNFB_SILENCE_MODULAR_DEPRECATION_WARNINGS = false;
-    });
-
-    describe('CustomProvider', function () {
-      if (!Platform.other) {
-        return;
-      }
-
-      it('should throw an error if no provider options are defined', function () {
-        try {
-          new firebase.appCheck.CustomProvider();
-          return Promise.reject(new Error('Did not throw an error.'));
-        } catch (e) {
-          e.message.should.containEql('no provider options defined');
-          return Promise.resolve();
-        }
-      });
-
-      it('should throw an error if no getToken function is defined', function () {
-        try {
-          new firebase.appCheck.CustomProvider({});
-          return Promise.reject(new Error('Did not throw an error.'));
-        } catch (e) {
-          e.message.should.containEql('no getToken function defined');
-          return Promise.resolve();
-        }
-      });
-
-      it('should return a token from a custom provider', async function () {
-        const spy = sinon.spy();
-        const provider = new firebase.appCheck.CustomProvider({
-          getToken() {
-            spy();
-            return FirebaseHelpers.fetchAppCheckToken();
-          },
-        });
-
-        // Call from the provider directly.
-        const { token, expireTimeMillis } = await provider.getToken();
-        spy.should.be.calledOnce();
-        token.should.be.a.String();
-        expireTimeMillis.should.be.a.Number();
-
-        // Call from the app check instance.
-        await firebase
-          .appCheck()
-          .initializeAppCheck({ provider, isTokenAutoRefreshEnabled: false });
-        const { token: tokenFromAppCheck } = await firebase.appCheck().getToken(true);
-        tokenFromAppCheck.should.be.a.String();
-
-        // Confirm that app check used the custom provider getToken function.
-        spy.should.be.calledTwice();
-      });
-
-      // TODO flakey after many runs, sometimes fails on android & ios,
-      // possibly a rate limiting issue on the server side
-      it('should return a limited use token from a custom provider', async function () {
-        const provider = new firebase.appCheck.CustomProvider({
-          getToken() {
-            return FirebaseHelpers.fetchAppCheckToken();
-          },
-        });
-
-        await firebase
-          .appCheck()
-          .initializeAppCheck({ provider, isTokenAutoRefreshEnabled: false });
-        const { token: tokenFromAppCheck } = await firebase.appCheck().getLimitedUseToken();
-        tokenFromAppCheck.should.be.a.String();
-      });
-
-      it('should listen for token changes', async function () {
-        const provider = new firebase.appCheck.CustomProvider({
-          getToken() {
-            return FirebaseHelpers.fetchAppCheckToken();
-          },
-        });
-
-        await firebase
-          .appCheck()
-          .initializeAppCheck({ provider, isTokenAutoRefreshEnabled: false });
-        const unsubscribe = firebase.appCheck().onTokenChanged(_ => {
-          // TODO - improve testing cloud function to allow us to return tokens with low expiry
-        });
-
-        // TODO - improve testing cloud function to allow us to return tokens with low expiry
-        // result.should.be.an.Object();
-        // const { token, expireTimeMillis } = result;
-        // token.should.be.a.String();
-        // expireTimeMillis.should.be.a.Number();
-        unsubscribe();
-      });
-    });
-
-    describe('AppCheck API methods', function () {
-      beforeEach(function () {
-        const { ReactNativeFirebaseAppCheckProvider } = appCheckModular;
-        let provider;
-
-        if (!Platform.other) {
-          provider = firebase.appCheck().newReactNativeFirebaseAppCheckProvider();
-          provider.configure({
-            android: {
-              provider: 'debug',
-              debugToken: getRandomToken(),
-            },
-            apple: {
-              provider: 'debug',
-              debugToken: getRandomToken(),
-            },
-            web: {
-              provider: 'debug',
-              siteKey: 'none',
-            },
-          });
-        } else {
-          provider = new ReactNativeFirebaseAppCheckProvider({
-            getToken() {
-              return FirebaseHelpers.fetchAppCheckToken();
-            },
-          });
-        }
-
-        // Our tests configure a debug provider with shared secret so we should get a valid token
-        firebase.appCheck().initializeAppCheck({ provider, isTokenAutoRefreshEnabled: false });
-      });
-
-      describe('config', function () {
-        // This depends on firebase.json containing a false value for token auto refresh, we
-        // verify here that it was carried in to the Info.plist correctly
-        // it relies on token auto refresh being left false for local tests where the app is reused, since it is persistent
-        // but in CI it's fresh every time and would be true if not overridden in Info.plist
-        it('should configure token auto refresh in Info.plist on ios', async function () {
-          if (Platform.ios) {
-            const tokenRefresh =
-              await NativeModules.RNFBAppCheckModule.isTokenAutoRefreshEnabled('[DEFAULT]');
-            tokenRefresh.should.equal(false);
-          } else {
-            this.skip();
-          }
-        });
-      });
-
-      describe('setTokenAutoRefresh())', function () {
-        it('should set token refresh', async function () {
-          firebase.appCheck().setTokenAutoRefreshEnabled(false);
-
-          // Only iOS lets us assert on this unfortunately, other platforms have no accessor
-          if (Platform.ios) {
-            let tokenRefresh =
-              await NativeModules.RNFBAppCheckModule.isTokenAutoRefreshEnabled('[DEFAULT]');
-            tokenRefresh.should.equal(false);
-          }
-          firebase.appCheck().setTokenAutoRefreshEnabled(true);
-          // Only iOS lets us assert on this unfortunately, other platforms have no accessor
-          if (Platform.ios) {
-            tokenRefresh =
-              await NativeModules.RNFBAppCheckModule.isTokenAutoRefreshEnabled('[DEFAULT]');
-            tokenRefresh.should.equal(true);
-          }
-        });
-      });
-
-      // TODO flakey after many runs, sometimes fails on android & ios,
-      // possibly a rate limiting issue on the server side
-      describe('getToken()', function () {
-        it('token fetch attempt with configured debug token should work', async function () {
-          try {
-            const { token } = await firebase.appCheck().getToken(true);
-            token.should.not.equal('');
-            const decodedToken = decodeJWT(token);
-            decodedToken.aud[1].should.equal('projects/react-native-firebase-testing');
-            if (decodedToken.exp * 1000 < Date.now()) {
-              return Promise.reject(
-                `Token expired? now ${Date.now()} token exp: ${decodedToken.exp * 1000}`,
-              );
-            }
-
-            // on android if you move too fast, you may not get a fresh token
-            await Utils.sleep(2000);
-
-            // Force refresh should get a different token?
-            // TODO sometimes fails on android https://github.com/firebase/firebase-android-sdk/issues/2954
-            const { token: token2 } = await firebase.appCheck().getToken(true);
-            token2.should.not.equal('');
-            const decodedToken2 = decodeJWT(token2);
-            decodedToken2.aud[1].should.equal('projects/react-native-firebase-testing');
-            if (decodedToken2.exp * 1000 < Date.now()) {
-              return Promise.reject(
-                `Token expired? now ${Date.now()} token exp: ${decodedToken2.exp * 1000}`,
-              );
-            }
-            (token === token2).should.be.false();
-          } catch (e) {
-            // we will silently pass rate limiting errors
-            e.message.should.containEql('Quota exceeded');
-            this.skip();
-          }
-        });
-
-        it('token fetch with config switch to invalid then valid should fail then work', async function () {
-          if (Platform.other) {
-            this.skip();
-          }
-          rnfbProvider = firebase.appCheck().newReactNativeFirebaseAppCheckProvider();
-          rnfbProvider.configure({
-            android: {
-              provider: 'playIntegrity',
-            },
-            apple: {
-              provider: 'appAttest',
-            },
-            web: {
-              provider: 'debug',
-              siteKey: 'none',
-            },
-          });
-
-          // Our tests configure a debug provider with shared secret so we should get a valid token
-          firebase
-            .appCheck()
-            .initializeAppCheck({ provider: rnfbProvider, isTokenAutoRefreshEnabled: false });
-          try {
-            await firebase.appCheck().getToken(true);
-            return Promise.reject(new Error('should have thrown an error'));
-          } catch (e) {
-            e.message.should.containEql('appCheck/token-error');
-          }
-
-          rnfbProvider.configure({
-            android: {
-              provider: 'debug',
-              debugToken: getRandomToken(),
-            },
-            apple: {
-              provider: 'debug',
-              debugToken: getRandomToken(),
-            },
-            web: {
-              provider: 'debug',
-              siteKey: 'none',
-            },
-          });
-          firebase
-            .appCheck()
-            .initializeAppCheck({ provider: rnfbProvider, isTokenAutoRefreshEnabled: false });
-
-          try {
-            const { token } = await firebase.appCheck().getToken(true);
-            token.should.not.equal('');
-            const decodedToken = decodeJWT(token);
-            decodedToken.aud[1].should.equal('projects/react-native-firebase-testing');
-            if (decodedToken.exp * 1000 < Date.now()) {
-              return Promise.reject(
-                `Token expired? now ${Date.now()} token exp: ${decodedToken.exp * 1000}`,
-              );
-            }
-          } catch (e) {
-            // we will silently pass rate limiting errors
-            e.message.should.containEql('Quota exceeded');
-            this.skip();
-          }
-        });
-      });
-
-      describe('getLimitedUseToken()', function () {
-        it('limited use token fetch attempt with configured debug token should work', async function () {
-          try {
-            const { token } = await firebase.appCheck().getLimitedUseToken();
-            token.should.not.equal('');
-            const decodedToken = decodeJWT(token);
-            decodedToken.aud[1].should.equal('projects/react-native-firebase-testing');
-            if (decodedToken.exp * 1000 < Date.now()) {
-              return Promise.reject(
-                `Token expired? now ${Date.now()} token exp: ${decodedToken.exp * 1000}`,
-              );
-            }
-          } catch (e) {
-            // we will silently pass rate limiting errors
-            e.message.should.containEql('Quota exceeded');
-            this.skip();
-          }
-        });
-      });
-
-      describe('activate())', function () {
-        if (Platform.other) {
-          return;
-        }
-
-        it('should activate with default provider and defined token refresh', function () {
-          firebase
-            .appCheck()
-            .activate('ignored', false)
-            .then(value => expect(value).toBe(undefined))
-            .catch(e => new Error('app-check activate failed? ' + e));
-        });
-
-        it('should error if activate gets no parameters', async function () {
-          try {
-            firebase.appCheck().activate();
-            return Promise.reject(new Error('should have thrown an error'));
-          } catch (e) {
-            e.message.should.containEql('siteKeyOrProvider must be a string value');
-          }
-        });
-
-        it('should not error if activate called with no token refresh value', async function () {
-          try {
-            firebase.appCheck().activate('ignored');
-            return Promise.resolve(true);
-          } catch (e) {
-            return Promise.reject(new Error('should not have thrown an error - ' + e));
-          }
-        });
-      });
-    });
-  });
-
   describe('modular', function () {
     let appCheckInstance;
