+++ packages/app/lib/internal/constants.ts
@@ -20,7 +20,6 @@ export const APP_NATIVE_MODULE = 'RNFBAppModule';
 export const DEFAULT_APP_NAME = '[DEFAULT]';
 
 export const KNOWN_NAMESPACES = [
-  'appCheck',
   'auth',
   'analytics',
   'remoteConfig',
