+++ packages/vertexai/lib/public-types.ts
@@ -16,7 +16,7 @@
  */
 import { AI, AIErrorCode } from '@react-native-firebase/ai';
 import { FirebaseAuthTypes } from '@react-native-firebase/auth';
-import { FirebaseAppCheckTypes } from '@react-native-firebase/app-check';
+import type { AppCheck } from '@react-native-firebase/app-check';
 
 /**
  * @deprecated Use the new {@link AI | AI} instead. The Vertex AI in Firebase SDK has been
@@ -33,7 +33,7 @@ export type VertexAI = AI;
  */
 export interface VertexAIOptions {
   location?: string;
-  appCheck?: FirebaseAppCheckTypes.Module | null;
+  appCheck?: AppCheck | null;
   auth?: FirebaseAuthTypes.Module | null;
 }
